raspberry-pi/requirements.txt
RPi.GPIO
requests
python-dotenv
